Escaping user defined data
Found a bug that caused bad url requests when you used values
with spaces.

Example: A level name of "My Awesome Level" would cause a url error because the spaces were not escaped.

Also changed the version number to 4.9.1 in hopes that it fixes the new user issue.
Please enter the commit message for your changes. Lines starting